Refactors in memory signature storage & documents
